- patricia.c:ascii2prefix() fixed compile warning
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid/trunk@7842 82007160-df01-0410-b94d-b575c5fd34c7

+++ src/patricia.c
@@ -140,7 +140,7 @@ New_Prefix (int family, void *dest, int bitlen)
 static prefix_t *
 ascii2prefix(int family, const char *string)
 {
-  unsigned long bitlen, maxbitlen = 0;
+  int bitlen, maxbitlen = 0;
   char *cp;
   char save[MAXLINE];
   union
@@ -167,7 +167,7 @@ ascii2prefix(int family, const char *string)
 
   if ((cp = strchr(string, '/')))
   {
-    bitlen = atol(cp + 1);
+    bitlen = atoi(cp + 1);
 
     /* *cp = '\0'; */
     /* copy the string to save. Avoid destroying the string */
